Adding communications and gearman files
The communication.js file defines communication between the job server, and
either the worker or the client. It also defines the packet types (byte magic
code, and byte type).

The gearman.js file is the main entry point in the package, and should
be used to create workers and clients, and do any interactions that are
non-specific to a worker or a client.